+++ core-ng/src/main/java/core/framework/impl/validate/BeanValidator.java
@@ -3,6 +3,6 @@ package core.framework.impl.validate;
 /**
  * @author neo
  */
-public interface ObjectValidator {
+public interface BeanValidator {
     void validate(Object instance, ValidationErrors errors, boolean partial);
 }