+++ core-ng/src/test/java/core/framework/impl/validate/BeanValidatorNotEmptyTest.java
@@ -14,12 +14,12 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 /**
  * @author neo
  */
-class ObjectValidatorNotEmptyTest {
-    private ObjectValidator validator;
+class BeanValidatorNotEmptyTest {
+    private BeanValidator validator;
 
     @BeforeEach
     void createObjectValidator() {
-        validator = new ObjectValidatorBuilder(Bean.class, Field::getName).build().get();
+        validator = new BeanValidatorBuilder(Bean.class, Field::getName).build().get();
     }
 
     @Test