+++ core-ng/src/test/java/core/framework/impl/web/bean/QueryParamBeanClassValidatorTest.java
@@ -1,0 +1,13 @@
+package core.framework.impl.web.bean;
+
+import org.junit.jupiter.api.Test;
+
+/**
+ * @author neo
+ */
+class QueryParamBeanClassValidatorTest {
+    @Test
+    void validate() {
+        new QueryParamBeanClassValidator(TestQueryParamBean.class, new BeanClassNameValidator()).validate();
+    }
+}