+++ core-ng/src/test/resources/validator-test/validator-notnull.java
@@ -1,26 +1,26 @@
-public class ObjectValidatorNotNullTest$Bean$ObjectValidator implements core.framework.impl.validate.ObjectValidator {
-    private void validateChildBean1(core.framework.impl.validate.ObjectValidatorNotNullTest.ChildBean bean, core.framework.impl.validate.ValidationErrors errors, boolean partial) {
+public class BeanValidatorNotNullTest$Bean$Validator implements core.framework.impl.validate.BeanValidator {
+    private void validateChildBean1(core.framework.impl.validate.BeanValidatorNotNullTest.ChildBean bean, core.framework.impl.validate.ValidationErrors errors, boolean partial) {
         if (bean.intField == null) {
             if (!partial) errors.add("child.intField", "intField must not be null");
         } else {
         }
     }
 
-    private void validateChildBean2(core.framework.impl.validate.ObjectValidatorNotNullTest.ChildBean bean, core.framework.impl.validate.ValidationErrors errors, boolean partial) {
+    private void validateChildBean2(core.framework.impl.validate.BeanValidatorNotNullTest.ChildBean bean, core.framework.impl.validate.ValidationErrors errors, boolean partial) {
         if (bean.intField == null) {
             if (!partial) errors.add("children.intField", "intField must not be null");
         } else {
         }
     }
 
-    private void validateChildBean3(core.framework.impl.validate.ObjectValidatorNotNullTest.ChildBean bean, core.framework.impl.validate.ValidationErrors errors, boolean partial) {
+    private void validateChildBean3(core.framework.impl.validate.BeanValidatorNotNullTest.ChildBean bean, core.framework.impl.validate.ValidationErrors errors, boolean partial) {
         if (bean.intField == null) {
             if (!partial) errors.add("childMap.intField", "intField must not be null");
         } else {
         }
     }
 
-    private void validateBean0(core.framework.impl.validate.ObjectValidatorNotNullTest.Bean bean, core.framework.impl.validate.ValidationErrors errors, boolean partial) {
+    private void validateBean0(core.framework.impl.validate.BeanValidatorNotNullTest.Bean bean, core.framework.impl.validate.ValidationErrors errors, boolean partial) {
         if (bean.stringField == null) {
             if (!partial) errors.add("stringField", "stringField must not be null");
         } else {
@@ -37,7 +37,7 @@ public class ObjectValidatorNotNullTest$Bean$ObjectValidator implements core.fra
         if (bean.children == null) {
         } else {
             for (java.util.Iterator iterator = bean.children.iterator(); iterator.hasNext(); ) {
-                core.framework.impl.validate.ObjectValidatorNotNullTest.ChildBean value = (core.framework.impl.validate.ObjectValidatorNotNullTest.ChildBean) iterator.next();
+                core.framework.impl.validate.BeanValidatorNotNullTest.ChildBean value = (core.framework.impl.validate.BeanValidatorNotNullTest.ChildBean) iterator.next();
                 if (value != null) validateChildBean2(value, errors, partial);
             }
         }
@@ -46,14 +46,14 @@ public class ObjectValidatorNotNullTest$Bean$ObjectValidator implements core.fra
         } else {
             for (java.util.Iterator iterator = bean.childMap.entrySet().iterator(); iterator.hasNext(); ) {
                 java.util.Map.Entry entry = (java.util.Map.Entry) iterator.next();
-                core.framework.impl.validate.ObjectValidatorNotNullTest.ChildBean value = (core.framework.impl.validate.ObjectValidatorNotNullTest.ChildBean) entry.getValue();
+                core.framework.impl.validate.BeanValidatorNotNullTest.ChildBean value = (core.framework.impl.validate.BeanValidatorNotNullTest.ChildBean) entry.getValue();
                 if (value != null) validateChildBean3(value, errors, partial);
             }
         }
     }
 
     public void validate(Object instance, core.framework.impl.validate.ValidationErrors errors, boolean partial) {
-        validateBean0((core.framework.impl.validate.ObjectValidatorNotNullTest.Bean) instance, errors, partial);
+        validateBean0((core.framework.impl.validate.BeanValidatorNotNullTest.Bean) instance, errors, partial);
     }
 
 }